refactoring: 地図描画部分
