+++ public/Data.js
@@ -1,4 +1,4 @@
-import location from "./Location.js";
+import { getCurrentLocation } from "./Location.js";
 const db = firebase.firestore();
 const collectionName = "comments";
 
@@ -15,11 +15,14 @@ const postData = (data, callback) => {
 };
 
 const getData = async () => {
+  const location = await getCurrentLocation();
   return await db
     .collection(collectionName)
     .get()
     .then(querySnapshot => {
-      return querySnapshot;
+      return querySnapshot.map(doc => {
+        return doc.data();
+      });
     });
 };
 export { postData, getData };
